+++ libraries/botbuilder-middleware-activity-type/src/index.ts
@@ -1,0 +1,5 @@
+/**
+ * @module botbuildercommunity/middleware-activity-type
+ */
+
+export * from './activityType';