+++ libraries/botbuilder-middleware-watson-nlu/src/entity.ts
@@ -1,17 +1,26 @@
 import { Middleware, TurnContext, ActivityTypes } from 'botbuilder';
-import { Engine } from '@botbuildercommunity/middleware-engine-core';
+import { Engine, TextAnalysisMiddleware } from '@botbuildercommunity/middleware-engine-core';
+import * as nlup from 'ibm-watson/natural-language-understanding/v1.js';
 import { WatsonEngine } from './engine';
+import { RANKING } from './schema';
 
 /**
  * @module botbuildercommunity/middleware-watson-nlu
  */
 
-export class EntityExtraction implements Middleware {
+export class EntityExtraction extends TextAnalysisMiddleware implements Middleware {
     public engine: Engine;
-    constructor(public serviceKey: string, public endpoint: string, public options?: any) {
+    public serviceKey: string;
+    public endpoint: string;
+    public options: any;
+    public constructor(serviceKey: string, endpoint: string, options?: any) {
+        super();
+        this.serviceKey = serviceKey;
+        this.endpoint = endpoint;
+        this.options = options;
         this.engine = new WatsonEngine(serviceKey, endpoint, options);
     }
-    public async onTurn(context: TurnContext, next: () => Promise<void>) {
+    public async onTurn(context: TurnContext, next: () => Promise<void>): Promise<void> {
         if(context.activity.type === ActivityTypes.Message) {
             const input = {
                 documents: [
@@ -22,7 +31,7 @@ export class EntityExtraction implements Middleware {
                 ]
             };
             try {
-                const result = await this.engine.entities(input);
+                const result = await this.engine.entities(input, this.config);
                 const l = result.documents[0].entities;
                 context.turnState.set('textEntities', l);
             }
@@ -31,5 +40,27 @@ export class EntityExtraction implements Middleware {
             }
         }
         await next();
+    }
+    public static getEntities(entitiesResult: nlup.EntitiesResult[], type?: string): string[] {
+        if(type != null) {
+            return entitiesResult
+                .filter((e: nlup.EntitiesResult): boolean => e.type.toLocaleLowerCase() === type.toLocaleLowerCase())
+                .map((e: nlup.EntitiesResult): string  => e.text);
+        }
+        return entitiesResult.map((e: nlup.EntitiesResult): string  => e.text);
+    }
+    public static rankEntityKeys(entitiesResult: nlup.EntitiesResult[], ranking: RANKING = RANKING.RELEVANCE): string[] {
+        return EntityExtraction
+            .rankEntities(entitiesResult, ranking)
+            .map((e: nlup.EntitiesResult): string => e.text);
+    }
+    public static rankEntities(entitiesResult: nlup.EntitiesResult[], ranking: RANKING = RANKING.RELEVANCE): nlup.EntitiesResult[] {
+        return entitiesResult.sort((a: nlup.EntitiesResult, b: nlup.EntitiesResult): number => b[ranking] - a[ranking]);
+    }
+    public static topEntity(entitiesResult: nlup.EntitiesResult[], ranking: RANKING = RANKING.RELEVANCE): string {
+        return EntityExtraction.rankEntityKeys(entitiesResult, ranking)[0];
+    }
+    public static topEntityResult(entitiesResult: nlup.EntitiesResult[], ranking: RANKING = RANKING.RELEVANCE): nlup.EntitiesResult {
+        return EntityExtraction.rankEntities(entitiesResult, ranking)[0];
     }
 }